requirements.txt
amqp==2.5.2
awesome-slugify==1.6.5
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.41.0
django-mailgun==0.9.1
django-summernote==0.8.11.6
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
Unidecode==0.4.21
urllib3==1.25.8
vine==1.3.0
zipp==2.2.0
